# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Battery Data Intelligence Engine",
    "description": "Intelligence API backing the admin dashboard. Protected routes verify Clerk session JWTs via JWKS; see app/api/deps.py.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/chemistries": {
      "get": {
        "tags": [
          "chemistries"
        ],
        "summary": "List Chemistries",
        "operationId": "list_chemistries_api_v1_chemistries_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_BatteryChemistryRead_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/chemistries/{chemistry_id}/notes": {
      "get": {
        "tags": [
          "chemistries"
        ],
        "summary": "List Chemistry Notes",
        "operationId": "list_chemistry_notes_api_v1_chemistries__chemistry_id__notes_get",
        "parameters": [
          {
            "name": "chemistry_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Chemistry Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AnalystNoteRead"
                  },
                  "title": "Response List Chemistry Notes Api V1 Chemistries  Chemistry Id  Notes Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "chemistries"
        ],
        "summary": "Create Chemistry Note",
        "operationId": "create_chemistry_note_api_v1_chemistries__chemistry_id__notes_post",
        "parameters": [
          {
            "name": "chemistry_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Chemistry Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalystNoteCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalystNoteRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/chemistries/{chemistry_id}/verified": {
      "patch": {
        "tags": [
          "chemistries"
        ],
        "summary": "Set Chemistry Verified",
        "operationId": "set_chemistry_verified_api_v1_chemistries__chemistry_id__verified_patch",
        "parameters": [
          {
            "name": "chemistry_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Chemistry Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifiedUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerifiedResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AnalystNoteCreate": {
        "properties": {
          "note_type": {
            "type": "string",
            "enum": [
              "data_error",
              "missing_data",
              "outdated",
              "other"
            ],
            "title": "Note Type"
          },
          "note_text": {
            "type": "string",
            "maxLength": 4000,
            "minLength": 3,
            "title": "Note Text"
          }
        },
        "type": "object",
        "required": [
          "note_type",
          "note_text"
        ],
        "title": "AnalystNoteCreate",
        "description": "Request body for per-entity POST \u2026/notes endpoints.\n\nClients send only ``note_type`` and ``note_text``.  The route handler\ninjects ``entity_type`` and ``entity_id`` from the URL path before writing\nto ``analyst_notes``."
      },
      "AnalystNoteRead": {
        "properties": {
          "note_type": {
            "type": "string",
            "enum": [
              "data_error",
              "missing_data",
              "outdated",
              "other"
            ],
            "title": "Note Type"
          },
          "note_text": {
            "type": "string",
            "maxLength": 4000,
            "minLength": 3,
            "title": "Note Text"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "entity_type": {
            "type": "string",
            "enum": [
              "company",
              "company_material_exposure",
              "company_supply_relationship",
              "company_regulation_exposure",
              "company_vehicle_model",
              "material",
              "hs_code_material_mappings",
              "regulation",
              "risk_event",
              "facility",
              "battery_chemistry"
            ],
            "title": "Entity Type"
          },
          "entity_id": {
            "type": "string",
            "title": "Entity Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "note_type",
          "note_text",
          "id",
          "entity_type",
          "entity_id",
          "created_at",
          "updated_at"
        ],
        "title": "AnalystNoteRead"
      },
      "BatteryChemistryRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "current_market_share_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Market Share Pct"
          },
          "market_share_as_of_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Market Share As Of Date"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "verified": {
            "type": "boolean",
            "title": "Verified",
            "default": false
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "latest_risk_score": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ChemistryRiskScoreRead"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "slug",
          "name",
          "status",
          "is_active",
          "created_at",
          "updated_at"
        ],
        "title": "BatteryChemistryRead"
      },
      "ChemistryRiskScoreRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "as_of_date": {
            "type": "string",
            "format": "date",
            "title": "As Of Date"
          },
          "methodology_version": {
            "type": "string",
            "title": "Methodology Version"
          },
          "material_concentration_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Material Concentration Score"
          },
          "geopolitical_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Geopolitical Score"
          },
          "composite_risk_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Composite Risk Score"
          },
          "score_confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score Confidence"
          },
          "computed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Computed At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "as_of_date",
          "methodology_version",
          "computed_at"
        ],
        "title": "ChemistryRiskScoreRead"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PaginatedResponse_BatteryChemistryRead_": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/BatteryChemistryRead"
            },
            "type": "array",
            "title": "Data"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          }
        },
        "type": "object",
        "required": [
          "data",
          "total",
          "page",
          "limit"
        ],
        "title": "PaginatedResponse[BatteryChemistryRead]"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "VerifiedResponse": {
        "properties": {
          "verified": {
            "type": "boolean",
            "title": "Verified"
          }
        },
        "type": "object",
        "required": [
          "verified"
        ],
        "title": "VerifiedResponse",
        "description": "Minimal response returned by all PATCH \u2026/verified endpoints."
      },
      "VerifiedUpdate": {
        "properties": {
          "verified": {
            "type": "boolean",
            "title": "Verified"
          }
        },
        "type": "object",
        "required": [
          "verified"
        ],
        "title": "VerifiedUpdate",
        "description": "Request body for PATCH \u2026/verified endpoints."
      }
    }
  }
}